% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.decode.arg}
\alias{.decode.arg}
\title{Decode from base64 and deserialize from json if necessary}
\usage{
.decode.arg(some.thing, simplifyMatrix = FALSE)
}
\arguments{
\item{some.thing}{the thing to be decoded and deserialized from json if necessary}
}
\value{
the decoded and deserialized argument
}
\description{
Work around the restrictions imposed by the Opal server on function arguments
The Opal server is very picky as regards function arguments. The workaround is
to serialize and encode them on the client and strip the right padding.
}
\details{
It looks for the string 'base64' in the argument to determine if it's encoded
}
