% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestDSS.R
\name{.deidentify}
\alias{.deidentify}
\title{: Remove potentially disclosive information}
\usage{
.deidentify(forestObject)
}
\description{
: De-identify a randomForest object, and at the same time
  make the ouptut lighter to send to the client.
}
