% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.init}
\alias{.init}
\title{Initialize the environment}
\usage{
.init()
}
\description{
Set a number of defaults for some of the functions.
These are: 1) hidden.fields - fields that must show no information (patient identifiers for instance)
2) allowed.funcs - supplementary functions allowed in dssDeriveColumn (see the documetation for that function)
3) join.pivot.col - the default column to join (or pivot) by. In the absence of this default
the 'by.col' argument in dssJoin and dssPivot becomes mandatory
}
