% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partKde2d.R
\name{.kde2d}
\alias{.kde2d}
\title{Two-Dimensional Kernel Density Estimation
From package MASS}
\usage{
.kde2d(
  x,
  y,
  h,
  n = 128,
  lims = unlist(rangeDSS(deparse(substitute(x)), deparse(substitute(y))))
)
}
\description{
Two-Dimensional Kernel Density Estimation
From package MASS
}
