% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.lock.unlock}
\alias{.lock.unlock}
\title{locks or unlocks bindings in environments}
\usage{
.lock.unlock(what, func)
}
\arguments{
\item{what}{a list of  environments and their respective objects - the output of ls.all above}

\item{func}{a function, either lockBinding or unlockBinding}
}
\description{
helper function for dssSubset and dssPivot
}
