% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.ls.all}
\alias{.ls.all}
\title{returns all objects in all environments}
\usage{
.ls.all(start = ".GlobalEnv")
}
\arguments{
\item{start}{a character the environment name where to start (default .GlobalEnv)}
}
\value{
a list of environment names and the respective objects defined in each environment
}
\description{
helper function for dssSubset and dssPivot
}
