% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestDSS.R
\name{.predict}
\alias{.predict}
\title{Predict classification of new patients
  using the forests of our different data sources.}
\usage{
.predict(forests, testData)
}
\arguments{
\item{forests:}{a list of randomForest objects.}

\item{testData:}{[data frame] new data to classify using the forests.
It must have at least the columns in `expl_vars`.
(We want to predict the value of `dep_var` for it.)}
}
\value{
a vector of length `nrow(testData)`.
}
\description{
For classification, we sum the votes from all the forests,
}
