% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestDSS.R
\name{forestDSS}
\alias{forestDSS}
\title{Construct of a random forest based on training set `df`.}
\usage{
forestDSS(df, dep_var, expl_vars, nodesize = 5, ...)
}
\arguments{
\item{df:}{[data frame] The training dataset.}

\item{dep_var:}{[string] the response factor ("y"), i.e. the categories
that will be the leaves of the tree.}

\item{expl_vars:}{[vector[string]] the classification variables.}

\item{nodesize:}{minimal number of points in the terminal nodes.
We fix this high enough to ensure more privacy. Default is 1
for classification and 5 for regression.}

\item{...:}{other parameters to `randomForest`.}
}
\value{
a randomForest object.
}
\description{
Calls `randomForest::randomForest` with formula dep_var ~ expl_vars.
  The number of trees produced is proportional to the number of rows in the dataset,
  but with a minimum of 10 and a maximum of 150 trees.
}
