% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregDSS.R
\name{linregDSS}
\alias{linregDSS}
\title{Returns the quantities shared to the client to performed distributed linear regression.}
\usage{
linregDSS(df, dep_var, expl_vars)
}
\arguments{
\item{dep_var:}{[string] the column name for the dependent variable (y).}

\item{expl_vars:}{[vector of strings] the column names for the explanatory variables.}
}
\value{
a list.
}
\description{
Caclulates X'X and X'y as in the normal equation, i.e. the MLE for beta
    is {{ b = inv(X'X) X'Y }}. No need to calculate beta here, nor to share the matrix X.
    Also returns y'y  and the sample size
    in order to recompute the T statistics (and the p-values).
}
