% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partColMeans.R
\name{partColMeans}
\alias{partColMeans}
\title{returns colmeans for this dataset}
\usage{
partColMeans(x, na.rm = FALSE, collist = NULL)
}
\arguments{
\item{x}{a dataframe}

\item{na.rm}{logical remove NAs?}

\item{collist}{a vector, subset of columns to use}
}
\value{
a numeric vector which contains the minimum and the maximum
}
\description{
the client can use this function to calculate the global colmeans
}
