% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partCov.R
\name{partCov}
\alias{partCov}
\title{calculates the matrix of sums of products of distance to the global means,}
\usage{
partCov(x, means = NULL, collist = NULL)
}
\arguments{
\item{x}{a data frame of matrix}

\item{means}{a vector of global means (one for each dimension)}

\item{collist}{a vector, subset of columns to use}

\item{na.rm}{logical remove NAs?}
}
\value{
a matrix of the sums of products for all pairs of dimensions
}
\description{
used as a piece of the global covariance matrix
the client sums all these local matrices and divides them by N-1 -> cov matrix
}
