% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialData.R
\name{partialData}
\alias{partialData}
\title{execute data() in the server session}
\usage{
partialData(what, start = NULL, end = NULL, package = NULL)
}
\arguments{
\item{start, }{keep rows starting with this}

\item{end, }{keep rows until here}

\item{package}{a character, name of the package where to look for datasets}

\item{x}{a character, name of a data frame}
}
\value{
a numeric vector which contains the minimum and the maximum
}
\description{
load specified dataset
}
