% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partRange.R
\name{partRange}
\alias{partRange}
\title{returns the minimum and maximum of a numeric vector or numeric columns of a data frame}
\usage{
partRange(...)
}
\arguments{
\item{x}{a numerical or a data frame}
}
\value{
a numeric vector which contains the minimum and the maximum or a data frame with such
vectors as columns
}
\description{
this function is similar to R function \code{range} but instead to not return
the real minimum and maximum, the computed values are multiplied by a very small random number.
}
