% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSubset.R
\name{safeSubset}
\alias{safeSubset}
\title{subset a dataframe}
\usage{
safeSubset(df, row.filter = TRUE, col.filter = TRUE)
}
\arguments{
\item{df}{a dataframe}

\item{row.filter}{a character, the filter for rows (as in '[')}

\item{col.filter}{a character, the filter for columns (as in '[')}
}
\description{
takes potentially unsafe row and column filters and evals them in the context of the dataframe.
It takes the following precautions:
* it takes a snapshot of all the envirnoments before the execution and deletes any new objects it finds after the execution
* it locks all environments before the execution, so no existing object can be modified
}
