% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showInfo.R
\name{showInfo}
\alias{showInfo}
\title{Show the levels for all factors in a dataframe}
\usage{
showInfo(df, limit.levels = "FALSE")
}
\arguments{
\item{df}{a dataframe}
}
\value{
levels for all non blacklisted columns
}
\description{
It filters out  a blacklist of columns (the options 'hidden.fields' and 'hidden.fields.regexes')
}
