% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showInfo.R
\name{showLevels}
\alias{showLevels}
\title{Show the levels in a factor}
\usage{
showLevels(x, limit.levels = "FALSE")
}
\arguments{
\item{x, }{a factor}
}
\value{
levels
}
\description{
A safer replacement for levels(). It should replace the definition of the 'levels' method in opal.
}
