% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{suggestPivotFormula}
\alias{suggestPivotFormula}
\title{Try to find an appropriate formula to pass to reshape2::dcast}
\usage{
suggestPivotFormula(what, cols = NULL, by.col = NULL)
}
\arguments{
\item{what, }{a data.frame}

\item{cols, }{a vector of columns used in the formula. They must exist  in 'what' (the previous argument). If null, all the columns from 'what' will be used.}

\item{by.col}{typically the patient id, the column that will end up as key of the resulting 'wide' data.frame. In most cases this
can be left null as it's been set when the package was loaded}
}
\value{
a formula (char) that can be (modified and) passed further to the function 'widen'
}
\description{
The CDISC format can be cumbersome for analysis. The solution would be to transform the data to wide format
using the function dcast from the package reshape2. However, as the actual data is not visible, choosing the right
widening formula is challenging. This function applies some simple heuristics on the provided parameters and tries to guess
a possible formula. This will not be necessarily the best option (and often it will fail completely) - it can still provide
a starting point for further refining.
}
