% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{widen}
\alias{widen}
\title{Create a wide data.frame using reshape2::dcast}
\usage{
widen(
  what,
  measure,
  cols = NULL,
  formula = NULL,
  by.col = NULL,
  completeCases = FALSE,
  fun.aggregate = NULL
)
}
\arguments{
\item{what}{a  data.frame}

\item{measure}{name of the column which stores values, this goes straight into the parameter value.var in reshape2::dcast}

\item{cols}{a vector of columns used in the formula. They must exist in 'what'. If null, all the columns from 'what' will be used.}

\item{formula}{the formula to use by reshape2::dcast. If left null an attempt will be made to guess it with suggest.pivot.formula.}

\item{by.col}{typically the patient id, the column that will end up as key of the resulting 'wide' data.frame. In most cases this
can be left null as it's been set when the package was loaded}

\item{completeCases}{a logical, keep only complete cases?}
}
\value{
a data.frame in wide format
}
\description{
This function  transforms the data from long (ex: CDISC) to wide format
using the function dcast from the package reshape2.
}
