% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{.init}
\alias{.init}
\title{Initialize the environment}
\usage{
.init()
}
\description{
Set a number of defaults for some of the functions.
These are:

1) hidden.fields - fields that must show no information (patient identifiers for instance)

2) allowed.functions - supplementary functions allowed in dssDeriveColumn (see the documentation for that function)

3) join.pivot.col - the default column to join (or pivot) by. In the absence of this default
the 'by.col' argument in dssJoin and dssPivot becomes mandatory

4) forbidden.options - a number of opal administrative options that are not allowed to be changed by dssSetOption

In order to set/modify any of these options the node administrator must create an .Rprofile file
in the rserver conf directory (normally /var/lib/rserver/conf) and populate it with the necessary
R commands to set the above options (see example). Then the following line must be added to the file
/var/lib/rserver/conf/Rserv.conf:

source /var/lib/rserver/conf/.Rprofile
}
\examples{
# content of the /var/lib/rserver/conf/.Rprofile file that adds as.POSIXct among
# the allowed functions in dssDeriveColumn:
options(allowed.functions='as.POSIXct')


}
